speech_to_text_project/requirements.txt
fastapi
uvicorn
openai-whisper
pydub
ffmpeg-python
python-multipart


